refactor: split preview.rs into separate modules (Phase 5.2)
Split 1880-line preview.rs into focused modules:
- common.rs: Shared utilities (format_size, get_border_style, etc.)
- text.rs: TextPreview with syntax highlighting (StyledLine, StyledSegment)
- hex.rs: HexPreview for binary files
- archive.rs: ArchiveEntry, ArchivePreview for zip/tar.gz
- pdf.rs: PdfPreview using pdftoppm
- image.rs: ImagePreview with ratatui-image protocols
- video.rs: VideoPreview with thumbnail and metadata
- diff.rs: DiffPreview for git diffs
- custom.rs: CustomPreview for external commands
- directory.rs: DirectoryInfo preview
- mod.rs: Module organization and re-exports

All existing tests pass (346 tests).
Part of FileView v2.0 architecture improvements.
