+++ src/render/mod.rs
@@ -20,8 +20,8 @@ pub use preview::{
     is_pdf_file, is_tar_gz_file, is_text_file, render_archive_preview, render_custom_preview,
     render_diff_preview, render_directory_info, render_hex_preview, render_image_preview,
     render_pdf_preview, render_text_preview, render_video_preview, ArchiveEntry, ArchivePreview,
-    CustomPreview, DiffPreview, DirectoryInfo, HexPreview, ImagePreview, PdfPreview, TextPreview,
-    VideoPreview,
+    CustomPreview, DiffPreview, DirectoryInfo, HexPreview, ImagePreview, PdfPreview, StyledLine,
+    StyledSegment, TextPreview, VideoPreview,
 };
 pub use ratatui_image::picker::Picker;
 pub use ratatui_image::FontSize;
