+++ src/render/preview/directory.rs
@@ -1,0 +1,135 @@
+//! Directory info preview
+
+use std::path::Path;
+
+use ratatui::{
+    layout::Rect,
+    style::{Color, Modifier, Style},
+    text::{Line, Span},
+    widgets::{Block, Borders, Paragraph},
+    Frame,
+};
+
+use super::common::{calculate_dir_size, format_size, get_border_style};
+
+/// Directory information for preview
+#[derive(Debug, Clone)]
+pub struct DirectoryInfo {
+    /// Directory name
+    pub name: String,
+    /// Number of files
+    pub file_count: usize,
+    /// Number of subdirectories
+    pub dir_count: usize,
+    /// Number of hidden items
+    pub hidden_count: usize,
+    /// Total size in bytes
+    pub total_size: u64,
+}
+
+impl DirectoryInfo {
+    /// Compute directory info from path
+    pub fn from_path(path: &Path) -> anyhow::Result<Self> {
+        let name = path
+            .file_name()
+            .map(|n| n.to_string_lossy().to_string())
+            .unwrap_or_else(|| path.display().to_string());
+
+        let mut file_count = 0;
+        let mut dir_count = 0;
+        let mut hidden_count = 0;
+        let mut total_size = 0u64;
+
+        if let Ok(entries) = std::fs::read_dir(path) {
+            for entry in entries.flatten() {
+                let entry_name = entry.file_name().to_string_lossy().to_string();
+                let is_hidden = entry_name.starts_with('.');
+
+                if is_hidden {
+                    hidden_count += 1;
+                }
+
+                if let Ok(metadata) = entry.metadata() {
+                    if metadata.is_file() {
+                        file_count += 1;
+                        total_size += metadata.len();
+                    } else if metadata.is_dir() {
+                        dir_count += 1;
+                        // Optionally calculate subdirectory size (can be slow for large dirs)
+                        if let Ok(dir_size) = calculate_dir_size(&entry.path()) {
+                            total_size += dir_size;
+                        }
+                    }
+                }
+            }
+        }
+
+        Ok(Self {
+            name,
+            file_count,
+            dir_count,
+            hidden_count,
+            total_size,
+        })
+    }
+}
+
+/// Render directory info preview
+pub fn render_directory_info(frame: &mut Frame, info: &DirectoryInfo, area: Rect, focused: bool) {
+    let separator = "─".repeat(area.width.saturating_sub(4) as usize);
+
+    let lines = vec![
+        Line::from(""),
+        Line::from(vec![Span::styled(
+            format!("  \u{f07b} {}", info.name), // Folder icon
+            Style::default()
+                .fg(Color::Blue)
+                .add_modifier(Modifier::BOLD),
+        )]),
+        Line::from(vec![Span::styled(
+            format!("  {}", separator),
+            Style::default().fg(Color::DarkGray),
+        )]),
+        Line::from(""),
+        Line::from(vec![
+            Span::styled("  Files:        ", Style::default().fg(Color::DarkGray)),
+            Span::styled(
+                format!("{}", info.file_count),
+                Style::default().fg(Color::White),
+            ),
+        ]),
+        Line::from(vec![
+            Span::styled("  Directories:  ", Style::default().fg(Color::DarkGray)),
+            Span::styled(
+                format!("{}", info.dir_count),
+                Style::default().fg(Color::Blue),
+            ),
+        ]),
+        Line::from(vec![
+            Span::styled("  Hidden:       ", Style::default().fg(Color::DarkGray)),
+            Span::styled(
+                format!("{}", info.hidden_count),
+                Style::default().fg(Color::Yellow),
+            ),
+        ]),
+        Line::from(""),
+        Line::from(vec![
+            Span::styled("  Total Size:   ", Style::default().fg(Color::DarkGray)),
+            Span::styled(
+                format_size(info.total_size),
+                Style::default()
+                    .fg(Color::Cyan)
+                    .add_modifier(Modifier::BOLD),
+            ),
+        ]),
+    ];
+
+    let widget = Paragraph::new(lines).block(
+        Block::default()
+            .borders(Borders::ALL)
+            .title(" Directory Info ")
+            .border_style(get_border_style(focused)),
+    );
+
+    frame.render_widget(widget, area);
+}
